refactor: impl_api_call_dot
