+++ state-chain/chains/src/dot/api/create_anonymous_vault.rs
@@ -1,9 +1,12 @@
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
 
-use crate::dot::{
-	Polkadot, PolkadotExtrinsicBuilder, PolkadotProxyType, PolkadotPublicKey,
-	PolkadotReplayProtection, PolkadotRuntimeCall, ProxyCall,
+use crate::{
+	dot::{
+		Polkadot, PolkadotExtrinsicBuilder, PolkadotProxyType, PolkadotPublicKey,
+		PolkadotReplayProtection, PolkadotRuntimeCall, ProxyCall,
+	},
+	impl_api_call_dot,
 };
 
 use crate::{ApiCall, ChainCrypto};
@@ -54,29 +57,7 @@ impl CreateAnonymousVault {
 	}
 }
 
-impl ApiCall<Polkadot> for CreateAnonymousVault {
-	fn threshold_signature_payload(&self) -> <Polkadot as ChainCrypto>::Payload {
-		self
-		.extrinsic_builder
-		.signature_payload
-		.clone()
-		.expect("This should never fail since the apicall created above with new_unsigned() ensures it exists")
-	}
-
-	fn signed(mut self, signature: &<Polkadot as ChainCrypto>::ThresholdSignature) -> Self {
-		self.extrinsic_builder
-			.insert_signature_and_get_signed_unchecked_extrinsic(signature.clone());
-		self
-	}
-
-	fn chain_encoded(&self) -> Vec<u8> {
-		self.extrinsic_builder.signed_extrinsic.clone().unwrap().encode()
-	}
-
-	fn is_signed(&self) -> bool {
-		self.extrinsic_builder.is_signed().unwrap_or(false)
-	}
-}
+impl_api_call_dot!(CreateAnonymousVault);
 
 #[cfg(test)]
 mod test_create_anonymous_vault {
