+++ state-chain/chains/src/dot/api/rotate_vault_proxy.rs
@@ -2,10 +2,13 @@ use codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use sp_std::{boxed::Box, vec};
 
-use crate::dot::{
-	BalancesCall, Polkadot, PolkadotAccountId, PolkadotAccountIdLookup, PolkadotExtrinsicBuilder,
-	PolkadotProxyType, PolkadotPublicKey, PolkadotReplayProtection, PolkadotRuntimeCall, ProxyCall,
-	UtilityCall,
+use crate::{
+	dot::{
+		BalancesCall, Polkadot, PolkadotAccountId, PolkadotAccountIdLookup,
+		PolkadotExtrinsicBuilder, PolkadotProxyType, PolkadotPublicKey, PolkadotReplayProtection,
+		PolkadotRuntimeCall, ProxyCall, UtilityCall,
+	},
+	impl_api_call_dot,
 };
 
 use crate::{ApiCall, ChainCrypto};
@@ -91,29 +94,7 @@ impl RotateVaultProxy {
 	}
 }
 
-impl ApiCall<Polkadot> for RotateVaultProxy {
-	fn threshold_signature_payload(&self) -> <Polkadot as ChainCrypto>::Payload {
-		self
-		.extrinsic_builder
-		.signature_payload
-		.clone()
-		.expect("This should never fail since the apicall created above with new_unsigned() ensures it exists")
-	}
-
-	fn signed(mut self, signature: &<Polkadot as ChainCrypto>::ThresholdSignature) -> Self {
-		self.extrinsic_builder
-			.insert_signature_and_get_signed_unchecked_extrinsic(signature.clone());
-		self
-	}
-
-	fn chain_encoded(&self) -> Vec<u8> {
-		self.extrinsic_builder.signed_extrinsic.clone().unwrap().encode()
-	}
-
-	fn is_signed(&self) -> bool {
-		self.extrinsic_builder.is_signed().unwrap_or(false)
-	}
-}
+impl_api_call_dot!(RotateVaultProxy);
 
 #[cfg(test)]
 mod test_rotate_vault_proxy {
