+++ state-chain/chains/src/eth/api/set_agg_key_with_agg_key.rs
@@ -2,7 +2,7 @@
 
 use crate::{
 	eth::{AggKey, Ethereum, SigData, Tokenizable},
-	ApiCall, ChainCrypto, impl_api_call_eth,
+	impl_api_call_eth, ApiCall, ChainCrypto,
 };
 
 use codec::{Decode, Encode, MaxEncodedLen};
