+++ state-chain/chains/src/eth/api/set_comm_key_with_agg_key.rs
@@ -1,6 +1,6 @@
 use crate::{
 	eth::{self, Ethereum, Tokenizable},
-	ApiCall, ChainCrypto, impl_api_call_eth,
+	impl_api_call_eth, ApiCall, ChainCrypto,
 };
 use codec::{Decode, Encode, MaxEncodedLen};
 use ethabi::{ParamType, Token};
